refactor: move the condition related to enableNativePlugin to the plugin file side
