+++ packages/vite/src/node/plugins/json.ts
@@ -7,9 +7,11 @@
  */
 
 import { dataToEsm, makeLegalIdentifier } from '@rollup/pluginutils'
+import { jsonPlugin as nativeJsonPlugin } from 'rolldown/experimental'
 import { SPECIAL_QUERY_RE } from '../constants'
 import type { Plugin } from '../plugin'
 import { stripBomTag } from '../utils'
+import type { ResolvedConfig } from '..'
 import { inlineRE, noInlineRE } from './asset'
 
 export interface JsonOptions {
@@ -37,10 +39,11 @@ const jsonLangRE = new RegExp(jsonLangs)
 export const isJSONRequest = (request: string): boolean =>
   jsonLangRE.test(request)
 
-export function jsonPlugin(
-  options: Required<JsonOptions>,
-  isBuild: boolean,
-): Plugin {
+export function jsonPlugin(config: ResolvedConfig, isBuild: boolean): Plugin {
+  const options = config.json
+  if (config.experimental.enableNativePlugin === true) {
+    return nativeJsonPlugin({ ...options, minify: isBuild })
+  }
   const plugin = {
     name: 'vite:json',
